RAU-78 Monkey patch is test case specific hence moved to test file and included one module in model for rails 4
