+++ lib/roles_on_routes/action_view_extensions/tags_with_roles.rb
@@ -41,25 +41,4 @@ module RolesOnRoutes
 end
 
 ActionView::Base.send :include, RolesOnRoutes::ActionViewExtensions::TagsWithRoles
-
-module ActionDispatch
-  module Routing
-    class RouteSet
-      def install_helpers(destinations = [ActionController::Base, ActionView::Base], regenerate_code = false)
-        Array(destinations).each { |d| d.module_eval {
-          include ActionView::Helpers;
-          include ActionDispatch::Routing::UrlFor
-          } }
-        named_routes.install(destinations, regenerate_code)
-      end
-      class NamedRouteCollection
-        def install(destinations = [ActionController::Base,ActionView::Base], regenerate = false)
-          reset! if regenerate
-          Array(destinations).each do |dest|
-            dest.__send__(:include, @module)
-          end
-        end
-      end
-    end
-  end
-end
+ActionView::Base.send :include, ActionDispatch::Routing::UrlFor
